requirements.txt
flask>=2.3.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
numpy>=1.24.0
requests>=2.31.0
